requirements.txt
moomoo-api>=1.0.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
apscheduler>=3.10.0
python-telegram-bot>=20.0
pytest>=7.4.0
pytest-mock>=3.11.0
